requirements.txt
django==4.2.11
python-decouple==3.8
django-phonenumber-field==7.3.0
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
Pillow==10.3.0
phonenumbers==8.13.35
drf-spectacular==0.27.2
django-parler-rest==2.2
django-admin-sortable2==2.1.10
django-ckeditor==6.7.1
django-mptt==0.16.0
uvicorn==0.29.0
psycopg2==2.9.9
django-cors-headers==4.3.1
django-modeladmin-reorder==0.3.1
requests==2.31.0
aiogram==3.6.0
loguru==0.7.2
pydantic_settings==2.2.1
fastapi==0.111.0
redis==5.0.4
eval_type_backport==0.2.0
django-better-admin-arrayfield==1.4.2
